+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Mapbox welcomes participation and contributions from everyone.
 * Add `MapboxMap.coordinatesForRect(rectF: RectF)` to support rectangle parameters.
 * Add `suspend` variants for the async `MapboxMap` functions : `queryRenderedFeatures`, `querySourceFeatures`, `setFeatureState`, `getFeatureState`, `removeFeatureState`, `getGeoJsonClusterLeaves`, `getGeoJsonClusterChildren`, `getGeoJsonClusterExpansionZoom`.
 * Add `MapboxMap.cameraChanges` returning Flow of camera updates.
+* Introduce custom lint rules to check illegal usage of literals in Expression DSL and suggest auto fix.
  
 ## Dependencies
 * Update dependencies.